OP cooldown

+++ modules/discord/api/v2/api_tomove.py
@@ -253,18 +253,18 @@ async def add_bot_command_api(request: Request, bot_id: int, command: PartialBot
     try:
         _tmp = enums.CommandType(command.cmd_type)
     except ValueError:
-        return ORJSONResponse({"done":  False, "reason": "UNSUPPORTED_MODE"}, status_code = 400)
+        return ORJSONResponse({"done":  False, "reason": "This command type is not yet supported", "code": 4658}, status_code = 400)
 
     id = await bot_auth(bot_id, Authorization)
     if id is None:
         return abort(401)
 
     if force_add is False:
-        check = await db.fetchrow("SELECT name FROM bot_commands WHERE name = $1 AND bot_id = $2", command.name, bot_id)
+        check = await db.fetchrow("SELECT name FROM bot_commands WHERE cmd_name = $1 AND bot_id = $2", command.cmd_name, bot_id)
         if check is not None:
-            return ORJSONResponse({"done":  False, "reason": "COMMAND_ALREADY_EXISTS"}, status_code = 400)
+            return ORJSONResponse({"done":  False, "reason": "A command with this name already exists on your bot", "code": 6858}, status_code = 400)
     id = uuid.uuid4()
-    await db.execute("INSERT INTO bot_commands (id, bot_id, cmd_groups, cmd_type, name, description, args, examples, premium_only, notes, doc_link) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)", id, bot_id, command.cmd_groups, command.cmd_type, command.name, command.description, command.args, command.examples, command.premium_only, command.notes, command.doc_link)
+    await db.execute("INSERT INTO bot_commands (id, bot_id, cmd_groups, cmd_type, cmd_name, description, args, examples, premium_only, notes, doc_link, friendly_name) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12)", id, bot_id, command.cmd_groups, command.cmd_type, command.cmd_name, command.description, command.args, command.examples, command.premium_only, command.notes, command.doc_link, command.friendly_name)
     await bot_add_event(bot_id, enums.APIEvents.command_add, {"user": None, "id": id})
     return ORJSONResponse({"done": True, "reason": None, "id": id, "code": 1001}, status_code = 206)
 
@@ -273,12 +273,12 @@ async def edit_bot_command_api(request: Request, bot_id: int, command: BotComman
     try:
         _tmp = enums.CommandType(command.cmd_type)
     except ValueError:
-        return ORJSONResponse({"done":  False, "reason": "UNSUPPORTED_MODE"}, status_code = 400)
+        return ORJSONResponse({"done":  False, "reason": "This command type is not yet supported", "code": 4658}, status_code = 400)
 
     id = await bot_auth(bot_id, Authorization)
     if id is None:
         return abort(401)
-    data = await db.fetchrow(f"SELECT id, cmd_type, cmd_groups, name, description, args, examples, premium_only, notes, doc_link FROM bot_commands WHERE id = $1 AND bot_id = $2", command.id, bot_id)
+    data = await db.fetchrow(f"SELECT id, cmd_type, cmd_groups, cmd_name, friendly_name, description, args, examples, premium_only, notes, doc_link FROM bot_commands WHERE id = $1 AND bot_id = $2", command.id, bot_id)
     if data is None:
         return abort(404)
 
@@ -287,7 +287,7 @@ async def edit_bot_command_api(request: Request, bot_id: int, command: BotComman
     for key in command_dict.keys():
         if command_dict[key] is None: 
             command_dict[key] = data[key]
-    await db.execute("UPDATE bot_commands SET cmd_type = $2, name = $3, description = $4, args = $5, examples = $6, premium_only = $7, notes = $8, doc_link = $9, cmd_groups = $10 WHERE id = $1", command_dict["id"], command_dict["cmd_type"], command_dict["name"], command_dict["description"], command_dict["args"], command_dict["examples"], command_dict["premium_only"], command_dict["notes"], command_dict["doc_link"], command_dict["cmd_groups"])
+    await db.execute("UPDATE bot_commands SET cmd_type = $2, cmd_name = $3, description = $4, args = $5, examples = $6, premium_only = $7, notes = $8, doc_link = $9, cmd_groups = $10, friendly_name = $11 WHERE id = $1", command_dict["id"], command_dict["cmd_type"], command_dict["cmd_name"], command_dict["description"], command_dict["args"], command_dict["examples"], command_dict["premium_only"], command_dict["notes"], command_dict["doc_link"], command_dict["cmd_groups"], command_dict["friendly_name"])
     await bot_add_event(bot_id, enums.APIEvents.command_edit, {"user": None, "id": command.id})
     return {"done": True, "reason": None, "code": 1000}
 
@@ -296,7 +296,16 @@ async def delete_bot_command_api_(request: Request, bot_id: int, command: BotCom
     id = await bot_auth(bot_id, Authorization)
     if id is None:
         return abort(401)
-    await db.execute("DELETE FROM bot_commands WHERE id = $1 AND bot_id = $2", command.id, bot_id)
+    if command.id:
+        cmd = await db.fetchval("SELECT id FROM bot_commands WHERE id = $1 AND bot_id = $2", command.id, bot_id)
+    elif command.cmd_name:
+        cmd = await db.fetchval("SELECT id FROM bot_commands WHERE cmd_name = $1 AND bot_id = $2", command.cmd_name, bot_id)
+    if not cmd:
+        return abort(404)
+    if command.id:
+        await db.execute("DELETE FROM bot_commands WHERE id = $1 AND bot_id = $2", command.id, bot_id)
+    elif command.cmd_name:
+        await db.execute("DELETE FROM bot_commands WHERE cmd_name = $1 AND bot_id = $2", command.cmd_name, bot_id)
     await bot_add_event(bot_id, enums.APIEvents.command_delete, {"user": None, "id": command.id})
     return {"done": True, "reason": None, "code": 1000}
